Refactor: Align architecture with pure OOP principles
This commit refactors the application to follow a rich domain model, fulfilling the request to make the architecture "perfectly object-oriented".

The key changes are:
-   **Anemic Domain Model to Rich Domain Model:** Business logic has been moved from service classes (`UserManagementService`) into the domain entities themselves (`Usuario`, `Persona`, `PoliticaSeguridad`).
-   **Introduction of Factories:** Created `PersonaFactory` and `UsuarioFactory` to encapsulate the complex logic of object creation and validation, cleaning up the service layer.
-   **Encapsulation of Update Logic:** Added `Update` methods to the domain entities, ensuring that objects are responsible for their own state changes. This removes direct property manipulation from the service layer.
-   **Simplified Service Layer:** The `UserManagementService` now acts as a lean coordinator, orchestrating calls to factories, repositories, and domain entities.
-   **Corrected Dependency Flow:** Fixed a build error caused by an incorrect dependency from `DataAccess` to `BusinessLogic`, ensuring the architectural layers remain cleanly separated.

The resulting architecture is more robust, maintainable, and strictly follows the principles of Object-Oriented Programming by ensuring data and behavior are encapsulated together.
